% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{clean_cci_vacc}
\alias{clean_cci_vacc}
\title{Clean CCI data, before standardizing}
\usage{
clean_cci_vacc(age_vacc_data_raw = cci_vacc_data)
}
\arguments{
\item{age_vacc_data_raw}{Raw data pulled from CCI github repo.}
}
\value{

}
\description{
Clean CCI data, before standardizing
}
\examples{

}
