% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{clean_vacc_age_cci}
\alias{clean_vacc_age_cci}
\title{Clean Vacc data}
\usage{
clean_vacc_age_cci(vacc_data, daily_state_vacc, state_pop_ageXyr)
}
\arguments{
\item{state_pop_ageXyr}{}
}
\value{

}
\description{
Clean Vacc data
}
