% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{convert_agegroups}
\alias{convert_agegroups}
\title{Function to transform population data from a worldpop age groups to user-defined age groups}
\usage{
convert_agegroups(age_pop_data, age_group_data, max_age = 100)
}
\arguments{
\item{age_pop_data}{data pulled from worldpop using `load_worldpop_age`}

\item{age_group_data}{}

\item{max_age}{max age to include (default is 100)}
}
\value{
long age population data by admin level 2, in user-defined age groups
}
\description{
Function to transform population data from a worldpop age groups to user-defined age groups
}
\references{
WorldPop (https://www.worldpop.org/geodata)
}
