% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{drop_dates_vaccdat}
\alias{drop_dates_vaccdat}
\title{Drop Problematic Dates from data}
\usage{
drop_dates_vaccdat(vacc_data_Xyr = vacc_data_10yr)
}
\arguments{
\item{vacc_data_Xyr}{}
}
\value{

}
\description{
Drop Problematic Dates from data
}
