% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{get_cci_vacc_Xyr}
\alias{get_cci_vacc_Xyr}
\title{Title}
\usage{
get_cci_vacc_Xyr(cci_vacc_clean, state_pop_ageXyr, daily_state_vacc)
}
\arguments{
\item{daily_state_vacc}{}
}
\value{

}
\description{
Title
}
\examples{

}
