% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{get_pop_Xyr_spline}
\alias{get_pop_Xyr_spline}
\title{Transform and distribute population to new age groups}
\usage{
get_pop_Xyr_spline(age_data, age_group_data)
}
\arguments{
\item{age_data}{# Data of population data of interest in some age groups. This should have column names: "location", "age", "age_l", "age_r", "pop"}

\item{age_group_data}{# Data on new age groups. This should have column names: "age", "age_l", "age_r"}
}
\value{

}
\description{
Transform and distribute population to new age groups
}
\examples{

}
