% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{get_standardized_us_agevacc}
\alias{get_standardized_us_agevacc}
\title{Title}
\usage{
get_standardized_us_agevacc(
  vacc_data = NULL,
  age_groups = "5yr",
  git_token = "AB63TGGKM6VBDNGJHXRRP7LAYPDIM"
)
}
\arguments{
\item{git_token}{}
}
\value{

}
\description{
Title
}
