% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{get_state_vacc}
\alias{get_state_vacc}
\title{Get Daily State Vaccination}
\usage{
get_state_vacc()
}
\value{
The output from this function is cleaned, up-to-date cumulative and daily COVID-19 vaccination numbers, by state, with dose 1 and 2 for Pfizer and Moderna vaccines (combined).
}
\description{
Get Daily State Vaccination
}
\examples{
get_state_vacc()

}
