% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{get_vacc_age_Xyr}
\alias{get_vacc_age_Xyr}
\title{Title}
\usage{
get_vacc_age_Xyr(
  vacc_clean = vacc_clean,
  state_pop_ageXyr = state_pop_ageXyr,
  daily_state_vacc = daily_state_vacc
)
}
\arguments{
\item{date_vacc_cum}{}
}
\value{

}
\description{
Title
}
