% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{get_vacc_Xyr_spline}
\alias{get_vacc_Xyr_spline}
\title{Title}
\usage{
get_vacc_Xyr_spline(
  vacc_age_data = vacc_clean,
  state_pop_ageXyr = state_pop_age5yr,
  daily_state_vacc
)
}
\arguments{
\item{state_pop_ageXyr}{}

\item{daily_state_vacc}{}
}
\value{

}
\description{
Title
}
