% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_state_curves}
\alias{plot_state_curves}
\title{plot_state_curves}
\usage{
plot_state_curves(
  vacc_data = vacc_data_10yr,
  states = NULL,
  nstates_sample = 12
)
}
\arguments{
\item{vacc_data}{}

\item{states}{}

\item{nstates_sample}{}
}
\value{

}
\description{
plot_state_curves
}
