% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{process_cci_age_vacc}
\alias{process_cci_age_vacc}
\title{Process CCI age vaccination data}
\usage{
process_cci_age_vacc(data = cci_vacc_data, daily_state_vacc, state_pop_ageXyr)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Process CCI age vaccination data
}
