% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{save_pdf_state_curves}
\alias{save_pdf_state_curves}
\title{save_pdf_state_curves}
\usage{
save_pdf_state_curves(
  vacc_data = vacc_data_10yr,
  pdf_filename = paste0("figures/prop_vacc_by_age10yr_", Sys.Date(), ".pdf")
)
}
\arguments{
\item{vacc_data}{}

\item{pdf_filename}{}
}
\value{

}
\description{
save_pdf_state_curves
}
