% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{transform_pop_agegroups}
\alias{transform_pop_agegroups}
\title{Title}
\usage{
transform_pop_agegroups(
  age_l_ = c(0, 12, 16, seq(25, 85, by = 10)),
  max_age = 100
)
}
\arguments{
\item{age_l_}{}

\item{max_age}{}
}
\value{

}
\description{
Title
}
