% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{get_clean_us_agevacc}
\alias{get_clean_us_agevacc}
\title{Title}
\usage{
get_clean_us_agevacc(
  git_token = "AB63TGALF5P7EDIKSV46LIDBEAGDS",
  min_vacc_age = 5
)
}
\arguments{
\item{git_token}{}
}
\value{

}
\description{
Title
}
