% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacc_funcs.R
\name{pull_cci_vacc}
\alias{pull_cci_vacc}
\title{Title}
\usage{
pull_cci_vacc(git_token = "AB63TGCIYANL5EQ647HRXH3AYAWZ6")
}
\arguments{
\item{git_token}{Token required to access data}
}
\value{

}
\description{
Title
}
